requirements.txt
numpy
Pillow
torch
torchvision

pytorch-lightning
numpy
scikit-learn
opencv-python-headless
Pillow
tqdm
grad-cam
streamlit